cogs_no_subdiretorio/requirements.txt
discord=2.3.0
discord.py=2.3.0
python-dotenv=1.0.0